requirements.txt
pymongo[srv]
python-dotenv
pydantic
pydantic-core
streamlit
parlant-client
sentence_transformers
einops
lagom
parlant